refactor(web): rename reqlog to requestlog
Matches sibling requestid and the new full-words naming rule; no ad-hoc
abbreviations. Pre-v1 rename, no behavior change.
